Game Switcher: fix PID wait
Initial issue: don't get the same behavior when launching the game from the game_list view vs. the game_detail view.
It works better on the game_list side, because on the game_detail side it looks like it doesn't wait for the PID to turn off: it goes right back to the Activities GUI, so the game runs at the same time as the GUI.
Note: the initial code from Nvrdid was working correctly